FindBugs code cleanups (3/13) - Buffer.
Buffer: make the BufferInfo member final. Enforces the implied invariant that the buffer info is initialized in the constructor and the reference is not modified again afterward. In practice this may have some (certainly not measurable) performance impact by aiding the compiler in optimization decisions, but this is mainly for code quality.