Fix SSL cert validation against built-in root CAs
Previously, the TrustManagerFactory was initialized with an empty
KeyStore, which caused it to reject all certificates.  Initializing
it with a null KeyStore instead will cause it to use the default
root CAs built in to Android, which will allow certificates issued
by those root CAs to be used without nagging the user.