fix reconnect when changing networks on andorid 5
They changed how the switching of networks happen in version 5, now if you transition from one network to another while both are available there will be a grace period before the OS shutsdown the old network, so the connection event on the new network happened before the disconnect from the old. Not explicitly disconnecting from the core if we get a broadcast that a network has been disconnected. (Android brodcasts that a network is disconnected a long time before it actually shuts the network down)